Rename package type to constants based on Ashik's comment
